text and file messaging, Google login, and styling
This commit implements the basic functionality for P2P text and file
messaging and also Google login using the Firebase Javascript SDK. Users
may see a list of all users who have signed up for the app, a list of
the conversations that have started, and the contents of those
conversations.

Styling is done with Material Design Lite.